[CORE][CONFIG] - Created new flood protection for Potion. - Removed flood protector sendMail. - Now FloodProtectorInterval can be a float value. - Fixed bug with augmentetion. - Format RequestRefine.java - Format FloodProtectors.java - Fixed bug with Geodata if target goes behind wall.

+++ trunk/gameserver/head-src/com/l2jfrozen/gameserver/util/FloodProtectorAction.java
@@ -46,7 +46,7 @@ public final class FloodProtectorAction
 	/**
 	 * Next game tick when new request is allowed.
 	 */
-	private volatile int _nextGameTick = GameTimeController.getGameTicks();
+	private volatile float _nextGameTick = GameTimeController.getGameTicks();
 	/**
 	 * Request counter.
 	 */
